+++ mycodo/mycodo_flask/utils/utils_input.py
@@ -156,6 +156,9 @@ def input_add(form_add):
         elif form_add.input_type.data == 'MH_Z16_I2C':
             new_sensor.location = '0x63'
             new_sensor.interface = 'I2C'
+        elif form_add.input_type.data == 'K30_I2C':
+            new_sensor.location = '0x68'
+            new_sensor.interface = 'I2C'
         elif form_add.input_type.data in ['K30_UART',
                                           'MH_Z16_UART',
                                           'MH_Z19_UART']: