+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Fix HDC1000 initialization issue ([#467](https://github.com/kizniche/mycodo/issues/467))
  - Fix Command PWM frontend issues ([#469](https://github.com/kizniche/mycodo/issues/469))
  - Fix ADC modules ([#482](https://github.com/kizniche/mycodo/issues/482))
+ - Update miflora to 0.4 ([#481](https://github.com/kizniche/mycodo/issues/481))
 
 
 ## 6.1.1 (2018-05-18)